2019/11/13

+++ src/main/java/com/jie/maven/firstspringboot/controller/PublishController.java
@@ -46,7 +46,7 @@ public class PublishController {
     public String dopublish(@RequestParam(value = "title", required = false) String title,
                             @RequestParam(value = "description", required = false) String description,
                             @RequestParam(value = "tag", required = false) String tag,
-                            @RequestParam(value="id",required = false) Integer id,
+                            @RequestParam(value="id",required = false) Long id,
                             HttpServletRequest request, Model model){
         model.addAttribute("title",title);
         model.addAttribute("description",description);
